store fragment m/z and intensities as float32, split frag_data columns
The one bytes-changing commit of the compaction (STORE_VERSION=3):

- frag_data (F,2) float64 -> frag_mz + frag_int float32 columns; the new
  names deliberately break any stale consumer of the packed array
- spectrum_mz / spectrum_int -> float32; all builders compute in float64
  and cast at storage; sort permutations run on the stored f32 values so
  stored order is self-consistent (parser, decoy builder, from_tagged)
- precursor-level scalars (prec_mz, prec_z, iRT, ion_mob) stay float64
- to_diann_df casts fragment columns to float64 at the Sage boundary;
  fragment_correlation/fragment_index already cast/quantize
- load() legacy fallbacks deleted (unreachable behind the version gate)
- transition verified: every f32 array within 4.8e-8 relative of the f64
  snapshot baselines (half an f32 ULP), all int/string fields byte-equal;
  snapshots regenerated; exact-equality frag asserts in test_mass_tags /
  test_iso_functions converted to f32-scale approx; full suite incl.
  integration green

~30 ppb m/z quantization vs ppm-scale tolerances; halves fragment-array
memory ahead of the single-copy layout change.

+++ src/iso_functions.py
@@ -305,8 +305,8 @@ def _alloc_iso_buffers(library, all_keys, n_iso):
 
     return (spec_offsets,
             spec_lengths,
-            np.empty(total, dtype=np.float64),
-            np.empty(total, dtype=np.float64),
+            np.empty(total, dtype=np.float32),
+            np.empty(total, dtype=np.float32),
             np.empty(total, dtype=np.int32))
